requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

apache-airflow-providers-databricks==6.10.0
apache-airflow-providers-google==10.23.0
pendulum
pandas
psycopg2-binary